requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
flasgger==0.8.0
Flask==0.12.2
flask-swagger==0.2.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
pymongo==3.6.0
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
gevent